django-app/requirements.txt
Django==5.1.2
crispy-bootstrap5
django-cors-headers
djangorestframework
data-wizard
Pillow